[Important] updating license
The project is now using the BSD 2-clause "simplified" license. There's no major difference between this and the old GPL v3 license other than better protection of the copyrights, we're only doing this change as the old license wasn't respected. Some people removed our credits or even replaced them with their own. That's not acceptable, this new license should be easier to understand and better in general for this project.