[Update] Added color codes to wanted level
Wanted level will be displayed in a more green color if it reduces fast,
white if violent and red if slow reduce or increasing. This is directly
connected to how cloose a law enforcer is so it's a way for criminals to
know if they should be worried or not.