[Update] Removed GTWgui
Although GTWgui made the windows look nicer it also removed some of
their features like move and resize, it also screwed up with the cursor
sometimes and forced a realod of all resources if a small change was
made to the GUI system, thus it's window feature is now removed.
However, GTWgui as a resource is still around to import custom fonts.